Moved SDK home synchronization from 'getExtraJvmArgs' to 'preImportCheck'.
Problem:
AndroidGradleProjectResolver#getExtraJvmArgs was meant to return any
extra JVM args to be passed to Gradle. In Android Studio, we needed to
do a "pre-import check", that is, make sure that both the IDE and
the project's local.properties (if present) point to the same Android
SDK home. The problem is that GradleProjectResolverExtension did not
provide any method where we can perform such check, and as a result,
we performed the check in 'getExtraJvmArgs' making this method have a
side effect.

Solution:
The nice JetBrains guys added the method 'preImportCheck' to help us
clean up our code. This CL takes advantage of this new method by
performing the SDK home sync there.

Change-Id: I2baad76dae512a9e3b48976b538fa3882793f660