Cut over to standard method for retrieving list of Activities from Manifest file
This is possible because of Siva's fix for the NPE that was being raised when
requesting ManifestInfo for the top-level module (which has no Android manifest file).

Also fix timing/threading issue in new code path where IntelliJ internals raised a
"Project already closed" exception on teh EDT - because the use of
SwingUtilities.invokeLater() (which is used to coalesce model sync events) creates the
possibility of attempting to retrieve the modules of a closed project.

Change-Id: Iaf522d982095073182c63aaae231bd75a64085dc