Don't throw exceptions while stopping gradle daemon
We attempt to stop all gradle daemons while exiting. This works by
locate the gradle executable and calling "gradle --stop". However,
if no projects are open, we cannot find the gradle executable and
in such a case, we throw an exception.

This CL changes the code to not throw an exception in such a case,
and just log a warning.

Change-Id: Ia90881099609b1aff855feca9c97237c8a0bb86c