More on managing relationship between Navigation Editor and Gradle build on IDE startup
This flaw shows itself when a Navigation Editor is left open when the IDE is closed.
On restarting the IDE, while the gradle build is taking place, certain files that are
required for correct rendering of layout files are in a state that confuses the Rendering
service - or at least leaves it unable to render 'correctly'. Normally, the symptoms
are partially rendered layout files - with missing backgrounds.

These effects are timing sensitive and so failures are intermittent.

The result of RenderService.getSession().getResult().isSuccess() is
true, both in the case of successful and unsuccessful renderings
(where, in this case, we take a rendering with a missing background
to be unsuccessful).

This CL revamps the internal event propagation mechanisms of the
Navigation Editor and associated classes to flush all state in the case
of a file system or rendering resource change (including, most importantly,
clearing any cached images of layout files) so that unsuccessful renderings
eventually get fully replaced by the successful ones that take place after the
gradle build process is complete.

Several optimisations are included in this CL to prevent cycles and performance
problems with this more thorough propagation mechanism.

Change-Id: Ibecf2306df8bfe540171dcccacb41790ed5fa70d