ResourceType inspection, and Magic Constant inspection for support annotations
This will flag two types of problems:

- Using the wrong resource type in parameters or returns from methods
  marked with one of the new resource type annotations in the support
  library (@StringRes, @DrawableRes, @ColorRes, and so on).

- Using the wrong constant in methods that expect an integer of String
  parameter to be one or more of a specific set of constants. This is
  similar to the existing MagicConstant inspection in IntelliJ, but
  works with the support library's @IntDef and @TypeDef annotations.

Code completion is also hooked up to this, so if you invoke code
completion on
   getResources().getString(
it will auto complete "R.string" (and import the application R class,
if necessary).

In an upcoming version of the Gradle plugin, AAR libraries will
package their source-retention annotations as annotations.zip files,
bundled next to the classes.jar and res resource folder. This CL looks
for annotations during Gradle sync and if found, registers these as
external annotations for the library. As a consequence, the above
inspections are run not just on platform code but against any library
(including the support library) which provides typedef and resource
annotations.

Change-Id: Ia4b1f7e24014d95e2e2a4c850955a3633f573ded