Improved handling of 'buildSrc' folder.
'buildSrc' is a special folder in Gradle that gets
compiled automatically. When an Android project has
this folder, the IDEA android plug-in is not
configuring the module corresponding to this folder
because the Gradle infrastructure in IDEA does not
let us know about the other modules in the project.

Until we have the chances we need in IDEA to decide
whether 'buildSrc' is part of an Android project,
we assume it is if the IDE is Android Studio. This
is a short-term workaround.

Also, this CL adds a change where we skip building
this folder explicitly, because it is handled
automatically by Gradle.

Change-Id: I91d15d1458dda3b85404613827604d8537917094