Now we find all variant selection conflicts, not just the first one.
For example, app1 and app2 depend on libraryX, but app1 depends on
libraryX variant 'A' while app2 depends on libraryX variant 'B'.
Since libraryX can have only one variant selected, a conflict is reported.

The report will show up in the "Messages" view, and it will contain
all the details about the conflict: the selected variant in libraryX
and what app1 and app2 expect. Double-clicking this report will
select the module (and scroll if necessary) in the "Build Variants"
view.

This work is necessary as a building block of the upcoming
"Project profiles" feature, where we would suggest users what modules
to import if this type of conflicts are detected in a project.

Change-Id: I6d2076b669e52429342c8d5ae25473e864543828