Fixes "SDK Manager" action, by getting the path from IDE's default SDK or project's local.properties.
Currently the "SDK Manager" action only works if the project has an Android SDK
configured. This is a limitation, because in case a Gradle sync fails, a Gradle-based
Android project may not have a SDK, and therefore this action won't run. This is
specially important when, in order to make a sync succeed, changes needed to be
made in SDK Manager (e.g. installing missing platforms.)

This CL fixes the action this way:

1. If we are running Android Studio and if it has a "default SDK" defined, we
   use the path of that SDK to open SDK Manager
2. If #1 fails, we use the SDK path in the project's local.properties file.
3. If #2 fails, we just fall back to the default behavior.

Change-Id: I7b9b36b082d33ae6332b6d34f37b693b038a21de