Support Source Set Folder Creation
This change adds support for the new templates which create source folders.
It adds a new constraint: source_set_folder which checks to see if the given
file path (relative to the module root) is a folder which is part of a source
set.

It also upgrades GradleFileMerger to handle dependency merging and literal substitution
which is important for renaming (e.g. manifest.srcFile 'AM.xml' -> manifest.srcFile 'AndroidManifest.xml')
Change-Id: Ib12bb604c55fcf34d35ef7021d05ff69637bdebf