Fix test failures.
This should fix two test failures:
1. Failure in org.jetbrains.android.AndroidProjectWizardTest.testCreateEmptyProject

This removes @NotNull annotation from a ctor parameter. In all fairness, I
believe it really belongs there. Parent project is all but required by the time
module is created. The code path in question seems to only be available through
a plugin installed in vanilla IntelliJ and I was not able to test it manually.

2. Failure in _LastInSuiteTest.testProjectLeak

This one was caused by AsyncValidatorTest not properly using the Disposer. This is
a "blind" fix as I don't know how to run this test suite locally.

Change-Id: I679d1bd4670b7078eae3d93d162a38a531c67aa7