Handle process cancelled exceptions from PSI infrastructure
During Lint iteration of the PSI, the PSI infrastructure
occasionally checks whether the current process has been
cancelled. This will happen if we for example start performing
editor inspections for the current file, and then the user
starts editing; at that point the current analysis is stale
so it cancels the operation. That means the PSI call will
abort with a ProcessCancelledException, and we need to
handle this gracefully such that we don't log a warning
that conversion failed; this is a normal scenario.

Also removed some (now) obsolete benchmarking code.

Change-Id: I6dd354e42ea1d42dc5906e7c97f21fce8ad983b2