Wait for emulator to become ready before installing apps
Studio assumes that as soon as a device is online, packages can
be installed on it. On emulators, adb may detect the device to be
online, but service like the package manager may not have been
started up as yet. Installing apps while in this state will result
in an inscrutable error: "Could not access the Package Manager".

This CL makes sure that the device is not just online, but has the
acore process running before attempting to install packages on it.

Fixes https://code.google.com/p/android/issues/detail?id=62113

Change-Id: I65319caedfaa2aff4dcdb3c30791ac624361755c