Show device chooser dialog if set of devices has changed.
The application launch configuration has a setting that allows users
to specify that the set of devices chosen in the device chooser dialog
should be reused for future launches. Once this is set, the only way
to get the device chooser dialog to show up again is to reset this
checkbox in the launch configuration.

This CL overrides that setting and shows the device chooser dialog
again if the list of device has changed from when the previous selection
was made. This is most likely to trigger in cases where a new device
is plugged in, or an existing device is unplugged.

In order to achieve this, each launch configuration now stores the
list of devices that were available at the time of the launch in addition
to the list of devices that were actually used. When the available list
doesn't match the previous available list, the device chooser is shown.

Fixes https://code.google.com/p/android/issues/detail?id=65231

Change-Id: I75ee897dbebde6cccc8a1c701ee9e7f913e60f69