Show banner when opening build.gradle in unimported project.
Banner will be shown in editor for build.gradle and settings.gradle if the
file is not located in the project or project module root. User can begin
import process by clicking the "Add Now..." link on the banner.

One potential issue is that this banner will always be shown in settings.gradle
file editors except for the files in the project root.

Change-Id: Ibfbb9d6f92cfb8c0c818c195b7d1d582f2f3c97b