Fixed bug in automatic variant conflict resolution on project sync.
The bug consisted in conflict resolution working under the
assumption than solving a conflict cannot create any additional
conflicts, which is not correct. For example, when having
2 modules depending on 2 different variants of the same module.

In this CL, after solving all conflicts on project sync we refresh
the list of conflicts and display the updated list. If there are
remaining conflicts, it means that the only way to solve them is
through "Project Profiles."

Change-Id: I11bc5661acd41f4a211ca4a551d60cd39a0e5816