Added "Module dependency info" to "Build Variants" view.
This CL adds an "information" button to each cell in the
"Build Variants" view where a module is displayed. This
"info" button shows a dialog where users can see:

- The modules (and variants) the selected module depends
  on.
- The modules (and variants) that depend on the selected
  module.

The idea is that build variants can get very complicated
(especially when having dependencies on specific variants
in library modules) and by showing this information we
can help users understand complex relationship between
modules.

This feature is not really meant for simple projects.

Change-Id: Ia75bf9cbb2e3f82a41b570566e534d9a365efac1