Add multiconfiguration preview for API levels
For various reason we use a single layoutlib (and a single set of loaded framework
resources), along with a special implementation of IAndroidTarget which reports an
older API level, but delegates everything else to the new layoutlib. This helps trick
the resource resolver into resolving best-fit resources from the given API level,
it helps the default themes chosen from layouts to reflect the preferred theme of
that API level (e.g. "Theme" for Gingerbread, "Theme.Holo" for IceCreamSandwich),
and so on.

It also adds id's to render preview, which lets you delete a particular preview (say
Nexus S from the device screen previews), and as you switch layouts it remembers
these deleted id's and excludes them there as well.

Change-Id: I7db2db3384f636d7ed26212213d6b02b1b996ac5