Cleanup methods used to determine if an app can be run on a device
AndroidFacet provided a couple of methods to determine whether the
current application could be run on a device or AVD. These methods
had a few issues, the primary issue being that in some cases, they
were checking the project target SDK against the device's API level
rather than the min sdk.

This CL moves this logic out to a separate class. The methods have
also been augmented to check for necessary features from the device.
(However, this feature is not used as yet: a future CL will pull
the list of required features from the manifest and pass them on
to the checker).

The device chooser dialog that displays the results of the
compatibility check has now been enhanced to display a reason why
a particular device has been determined to not be compatible if
that is the case.

Change-Id: Ifb0a0e2ce510da3f04278f81affa55348fc2b5be