Query Gradle model for whether the apk is debuggable.
Only apks that have the debuggable attribute set in their manifest
can be debugged. The android plugin already had a check for this, but
that check looked only at the manifest. This CL fixes the setup for
Gradle projects by querying the Gradle model when available.

Also fixes a minor issue: when a debug launch that embeds to an existing
session failed, it was leaving the process running. The process is now
properly terminated.

Change-Id: If07fbd60f95f6133b6394575e05253040a70b4d2