Fixed timeout of Gradle daemons.
By default IDEA gives Gradle daemons 1 minute to stay idle,
when the default is 3 hours. This setting causes Gradle to
create daemons pretty often, which has a lot of overhead.

The way IDEA sets the timeout is through a system property.
In this CL we override set the corresponding system
property to -1, so IDEA won't change the default value.

This system property only affects project sync. We were
setting the same default (1 minute) in project builds
(both JPS and GradleInvoker.) This CL fixes the daemon
timeout for project builds as well.

Change-Id: Ibf82655f8e053489dbf29d00c1532cfe06de56a1