Improve create resource dialog for invalid resource names
The create resource dialog, shown for example if you apply an editor quickfix
to create a layout resource file which doesn't exist, normally will hide the
filename text field if the file name is already known.

However, if the filename is invalid (for example, contains uppercase characters)
this is problematic, since the name validator will catch that something is wrong,
but you can't edit the filename.

Second, it would focus the second text field rather than the filename textfield
(and the second field has completion, which would cause a balloon tip which
obscured the erroneous filename).

Third, it would not trigger validation until the first textfield edit event.

(Related to issue 67671)

Change-Id: I8de1df51e3b8a11468fe2e1b45a8f00dd071e95f