Improve launch compatibility checks.
This CL extracts the list of required by the application from its
manifest and passes it on to the launch compatibility test. This allows
the device chooser to show that apks that use the watch feature can
only be run on a watch.

It also introduces a check in the other direction, specifically for
watch devices: If an apk does not specify that it can run on a watch,
then it cannot be installed on a watch.

Change-Id: Iefab322858926f0a7276d7d6d74deb9d30c3dfe7