Fixed project sync for new projects, when the IDE has open projects.
When creating a new project, and the IDE has already one or more
open projects, the IDE opens the new project, automatically triggering
a sync. To have full control of when syncing happens we set a
property in the project, but we did not do it for this use case.

This CL sets the property before the check for automatic sync happens.

Change-Id: I28a08c8d84af80d940dc2d3aa3568f59488d5ecd