Fixed bug where BuildExceptions were not properly handled.
If a project build failed, and the error messages were not
displayed in the "Messages" window, the IDE ignored completely
the BuildException thrown by Gradle.

In this CL, if there are no error messages displayed, but
there is a BuildException, we display an error message using
the output of stderr. Therefore there is at least one error
and the build is treated as "failed."

Change-Id: I27a9c37b203412bed7b663f97e09db9d035d15ae