Support templates.zip from Exploded AARs
In order to shipping template revisions quickly in a versioned, packaged fashion,
we now support bundling those templates in AAR files. This CL adds functions which
scan through the exploded-aars in a given project and search for a 'templates.zip'
file in the top level of each AAR. If found the file will be unzipped to a temp directory
and processed as a normal template container.

Finer points:

* ONLY templates from the highest version of each AAR will be included.
* The templates.zip file is ONLY unzipped once (for a given version of the AAR)

Change-Id: I90be8d175e19cfd1908c1a404b23d7f9851e5dbe