Replaced Navigatables with Hyperlinks in "Messages" items.
Before the "Messages" window supported hyperlinks, we used
Navigatables to, well, navigate in the IDE (e.g. to open SDK
manager, or find references of unresolved dependencies, etc.)

Now that the "Messages" window supports hyperlinks, it is
more intuitive for users to click them in order to fix issues
reported in the "Messages" window. This CL replaces all the
old Navigatables with hyperlinks.

Change-Id: I68bbc258a7b7eb823289e883e7e98d05f58f7029