Add intention action for forking resources
We've had an action for taking the current layout and forking it in
a different configuration. This CL generalizes this to (1) other file resources
and (2) value resources.

For example if you update your parent theme to Theme.Material.Light, lint
will flag it as requiring v21; this CL adds a quickfix to then override
the style in values-v21. It is also an intention action so you can for
example invoke it on a dimension and choose to override it in layout-xlarge,
and so on.

Change-Id: I25cc991fa8d46dbd048036a6940afe3fa56a4d41