Add API Level filtering for Form Factors
This change adds the ability to blacklist/whitelist api levels for
different formfactors.

If a whitelist is present, the blacklist will be ignored, and only
items on the whitelist will be displayed to the user as choices for
the minSDK level.

If no whitelist is present, items on the blacklist (if any) as well as any items
below the specified minimum value for the form factor will be removed from
the list.

This change also stores information about default API levels in Studio
rather than the templates (since the templates sometimes are missing
this info)

Change-Id: I6a0438547f6f7007f365ef8738278c395815355f