Fix compileSdkVersion to work with preview platforms in the PSD
The project structure dialog was putting the version codename into
the compileSdkVersion dropdown model instead of the platform hash.
(We need the platform hash for compileSdkVersion, and the codename
for minSdkVersion and/or targetSdkVersion).

It also fixes the jumbled order of targets by obtaining the targets
directly from sdklib instead of going through the SDKs mapped from
sdklib as IntelliJ SDK instances.

Change-Id: Ifd6b2157b0dbaaffcb963f2f6e373b9578b15975