Supports exclusion syntax in dependency statements
Improves dependency parsing for Project Structure such that it will
no longer stomp on exclude statements in closures attached to
dependencies. It doesn't display the excludes in the UI, but that
might be an opportunity for a future CL.

Also improves parsing so it won't choke on
compile(path: '...', configuration: '...') statements.

Bug: 68441
Change-Id: I69b6c43b91c7b3c643ffcc45bc3f85f9c77455b5