Dependency lookup dialog fixes
In the Maven dependency lookup dialog, fixes some issues with the
hardcoded list of dependencies. For dependencies coming from
the Android/Google support repositories, it uses the
RepositoryUrlManager to determine what libraries are present and
to fetch the latest installed version, taking into account whether
or not the module is using a preview platform to determine whether
to show preview libraries or not.

It now uses specific version numbers of dependencies instead of the
+ syntax for dynamic resolution.

It also now omits the pacakging from the coordinates it generates;
specifying the packaging is unnecessary and breaks transitive
dependency resolution in Gradle.

Bug: 72963
Bug: 71106
Change-Id: I4426982001ca012c8c586de0b59b3bad932eb9e5