Fix module import UI.
It is now possible to rename modules to import, including the
capability to use hierarchical Gradle names.
JTable was replaced with a custom UI to make the UI more detailed
and easier to navigate. Module-specific statuses are now always
shown, as opposed to tooltips used before.
Controller layer was refactored to make it more testable and to
handle edge cases better.

Change-Id: Ib7de2a5c662dc20470e773c914683762cbe64e3a