Search MavenCentral directly from Dependencies dialog
In Project Structure, in the maven dependency lookup dialog, search
directly against MavenCentral instead of using the local index cache
of it. For one, we would like to make the building of this cache
optional, and for another, indexing doesn't work for jcenter.
For that matter, this CL doesn't search against jcenter either, but
they don't appear to have a search UI that's useful to us that
doesn't require an authenticated user.

Bug: 72061
Change-Id: I3f6b2a79957b567b3c4cae09d6d38c004a3b29a1