Add action to view bitmaps from the debugger
This CL adds an action "View Bitmap" that is displayed when you
right click on a android.graphics.Bitmap object in the stack
frame while debugging. Selecting that action retrieves the bitmap
data corresponding to that object from the device, and displays
it in a popup.

Currently, the max size of the bitmap is restricted to 1024 pixels.
If the bitmap exceeds that size, then it is scaled down.

There are a few simplifying assumptions:
 - We assume that the bitmap object always has an "mBuffer" field
   that has the bitmap data accessible from the CPU. This is true
   for now, but might change in future platform version. A better
   approach might be to explicitly copy pixel data to a buffer and
   then retrieve the contents of that buffer.

 - We assume that the bitmap is of CONFIG_ARGB, and we assume that
   the buffer format is actually consistent across devices. This
   is supposedly a valid assumption, but it is not enforced anywhere.

Change-Id: I14fd5a42880ef991b8ad2233e714260ae3d903fb