navigator: Show instrumentation test sources
Android instrumentation test sources are provided as part of
an extra source set. This CL incorporates those source sets when
computing the list of files to the Android Navigator view.

Currently, the test sources are actually merged along with
the regular sources. In a future CL, we can investigate making
this an option (possibly pulling out the tests and all source types
to be a separate node).

Fixes https://code.google.com/p/android/issues/detail?id=74063

Change-Id: I93c7ef15af98e05c96bd15a6ed85f47bf843bc2b