Fixed bug where compiler errors where not shown by JPS builder.
This issue appears when the output parsers do not recognize the
output of stderr (this is the case when the error comes from
Gradle itself, and not a 3rd-party tool like javac.) The fix is
to create an error message from stderr, which contains the
cause of the failure.

This CL also updates UI test MultipleModuleTypeCompilationTest
to reproduce the issue and verify that is properly fixed. In addition,
IdeFrameFixture now guarantees that GradleInvoker (in-process
Gradle builds) is always enabled.

Change-Id: Iafbf4105e236e7f42f0f1df8a8f464824ac28131