device frames: Don't support framing images < artwork's screen size
The device art painter supports stretching the artwork (as long as
the aspect ratios are the same) to fit the image. But we don't support
the reverse case where the image is of lower size than the artwork.

With this change, we won't show device frames of newer devices that
happen to be of a same aspect ratio but higher density than older
low res devices.

Fixes https://code.google.com/p/android/issues/detail?id=74030

Change-Id: I20ae4da982cfefb361421a1b8544a44a4c6b480f