Now we detect and fix if the local Gradle distribution has an unsupported version.
The next version of the Android Gradle plug-in (0.13.0) will require
Gradle 2.0 as a minimum. This CL checks that the local Gradle distribution
used by a project is at least Gradle 2.0. If it is not, it prompts the
user to select a local Gradle 2.0 distribution and then continue with
Project sync.

This CL updates the UI test PluginAndGradleUpgradeTest to test this
scenario.

Change-Id: I3c80cd375f99b984cf22b9804b8fd46d5ef80b95