Fixes artifact that occurred when tooltips are shown.
When a tooltip is drawn over the canvas the repaint method
was called with a very small clipping area. Redrawing
that area only didn't change anything else, but it caused the timer
to be retriggered skipping a frame. While all the fade-in/out
animation of the tooltip was happening, the timeline frames were
being skipped.

Change-Id: I61c5a0e7541b031310209fad47dfa6bf3a5701e6