Handle API levels properly on configuration switch
If a layout was overriden in another configuration with a different API
level, switching configurations would lead to unexpected changes. This
is mainly because of the fact that we want to be able to render layouts
for older API levels with newer versions of the platform because of
various fidelity bugs with old LayoutLibs.

This change removes the VersionQualifier from the configuration before
finding the best match for the layout. If the current file is part of
this list of matching layouts, we don't switch.

Bug: http://b.android.com/75156
Change-Id: I9f4527c74ad8d9445ab812630e659c4c99f9589a