Provide a non blocking API to obtain adb bridge.
Currently, the way to get hold of adb connection is via
AndroidSdkData.getDebugBridge(). This method can only be called
from the EDT, and it is a blocking call that launches a modal
dialog while waiting for the adb connection to be established.

This CL introduces a new API: AdbService.initializeAndGetBridge(),
which returns a Future and is non blocking. This method
initializes the bridge, and launches a background task that
will set the future when the connection is established.

The existing API has been reworked to use this new method,
but provide the same behavior as before. So all callers of the
existing API will see the same behavior as before.

In a future CL, we'll rework some of the clients to not use
the blocking API and use the non blocking API instead.

Change-Id: I4a1c2fd5ca2b5e312175848761af4c8da7c0efe4