[UI Testing] Fixed bug where hyperlinks were not found in Messages window.
The problem is that, because the text of the hyperlink can be
spread in multiple lines, we were trying to find the hyperlink
to click on by matching the hyperlink against the expected
text, but using the expected text as a regular expression
pattern. This easily breaks if the expected text has regex
special characters like "(".

The solution is to do the opposite and put the text of a
hyperlink in a single line, and do a match by simply calling
"equals."

This CL includes cleanup:
- Added method openSimpleApplication to GuiTestCase (this was
  duplicated in many test cases)
- Added a MessageMatcher that matches the message (in the
  "Messages" tool window) by checking if the first line of
  the message starts with some expected text

Change-Id: I485bb42fc198aff12939e433d0b4754ccbbc8b95