ResFolderRepository: scanning files should hold a write lock
Here's a scenario where the current locking mechanism is incorrect:
 - A PsiFile is modified
 - On the EDT thread, ResourceFolderRepository's scanValueFile is
   called which is run under a read action
 - On a background thread, a method is iterating over items obtained as:
     List<ResourceItem> items = repository.getResourceItem(ResourceType.STRING, key);
   This background thread holds a read lock.

In this particular scenario, the background thread could get a
ConcurrentModificationException because the list returned by the resource repository
is being concurrently modified by scanValueFile in the EDT thread.

This CL changes ResourceFolderRepository to run these actions under a write
lock rather than a read lock. This could possibly be optimized to not hold
the write lock the entire time, but just during the time of the update of the
shared data structures.

Change-Id: I208cdb91301bbca088ad5043b32926789a95e951