Better message for sync error "unexpected end of block data".
This issue usually happens:
1. On Linux (we confirm that this does not happen on Mac or Windows)
2. When build-tools or Android platform is missing
3. Inside Android Studio during project sync (not when building the
   project - on the command line or IDE)

The error happens on the Gradle side, and it has been difficult to
debug and find the root cause. This CL does not fix the bug, but at
least offers a better message to the user. It also shows the
"Open Android SDK" and "Build Project" hyperlinks, next to the
error message.

https://code.google.com/p/android/issues/detail?id=72556

Change-Id: I8c28dda4996c9ed01757bc7dfa34e3c111a1b7a4