Add an Android-specific implicity property usage provider
If you run Analyze > Inspect Code... on a default Android project, you
end up with a dozen warnings for unused properties in .property
files. Some of these are for properties in local.properties, such as
sdk.dir and ndk.dir. These aren't really unused; they are read by the
Gradle plugin (and the IDE) when building the project, but the IDE
analysis doesn't know that, so they are marked as unused in the editor
and listed in the inspections output.

Similarly, all the properties in the Gradle wrapper property file
(gradle/wrapper/gradle-wrapper.properties) such as distributionUrl are
also all listed. These are also used; not by the gradle plugin this
time but the gradlew wrapper code.

The IDE has a facility for registering "implicit property usages", so
we use that mechanism to mark the above properties are implicitly
used, which removes the various unused warnings.

Change-Id: Id426d010062394739457c28b71703f5486481f97