[UI Testing] Fixed bug where FEST tried to click "Project Sync" button, but ended up clicking another button.
The problem was that even though FEST was waiting for sync and any background task
to finish before requesting a new sync, the "Run Configurations" drop-down in the IDE
toolbar was not populated yet, and as a result, the "Project Sync" button was not in
its final position.
The side effect was that FEST sometimes clicked the "Project Structure" button instead.

Change-Id: I14026afe8f6af96470eb2f0ee85c4f7b45121b75